add(ep-approval): new entry to split the records
